discard signal on non windows OS
