+++ versionmanager/proxy/light/transmit_others.go
@@ -1,0 +1,29 @@
+//go:build !windows
+
+/*
+ *
+ * Copyright 2024 tofuutils authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package lightproxy
+
+import "os"
+
+func transmitSignal(signalReceiver <-chan os.Signal, _ *os.Process) {
+	for range signalReceiver {
+		// discard signals on non Windows OS (already send to whole process group)
+	}
+}
