+++ versionmanager/proxy/proxy.go
@@ -29,19 +29,13 @@ import (
 	"github.com/hashicorp/hcl/v2/hclparse"
 
 	"github.com/tofuutils/tenv/v4/config"
-	"github.com/tofuutils/tenv/v4/config/envname"
-	cmdproxy "github.com/tofuutils/tenv/v4/pkg/cmdproxy"
+	"github.com/tofuutils/tenv/v4/pkg/cmdproxy"
 	"github.com/tofuutils/tenv/v4/pkg/loghelper"
 	"github.com/tofuutils/tenv/v4/versionmanager/builder"
 	"github.com/tofuutils/tenv/v4/versionmanager/lastuse"
-	detachproxy "github.com/tofuutils/tenv/v4/versionmanager/proxy/detach"
 )
 
-const (
-	chdirFlagPrefix = "-chdir="
-
-	msgReadDefaultDetachErr = "Failed to read " + envname.TenvDetachedProxyDefault + " environment variable, default to false :"
-)
+const chdirFlagPrefix = "-chdir="
 
 // Always call os.Exit.
 func Exec(conf *config.Config, builderFunc builder.Func, hclParser *hclparse.Parser, execName string, cmdArgs []string) {
@@ -66,13 +60,6 @@ func Exec(conf *config.Config, builderFunc builder.Func, hclParser *hclparse.Par
 	execPath := ExecPath(installPath, detectedVersion, execName, conf.Displayer)
 
 	cmd := exec.CommandContext(ctx, execPath, cmdArgs...)
-
-	defaultDetach, err := conf.Getenv.Bool(false, envname.TenvDetachedProxyDefault)
-	if err != nil {
-		fmt.Println(msgReadDefaultDetachErr, err) //nolint
-	}
-
-	detachproxy.InitBehaviorFromEnv(cmd, conf.Getenv, defaultDetach)
 
 	cmdproxy.Run(cmd, conf.GithubActions, conf.Getenv)
 }
